+++ crates/rust-client/src/lib.rs
@@ -118,7 +118,10 @@ pub struct Client<R: FeltRng> {
     rpc_api: Box<dyn NodeRpcClient + Send>,
     /// An instance of a [LocalTransactionProver] which will be the default prover for the client.
     tx_prover: Arc<LocalTransactionProver>,
+    /// An instance of a [TransactionExecutor] that will be used to execute transactions.
     tx_executor: TransactionExecutor,
+    /// Flag to enable the debug mode for scripts compilation and execution.
+    in_debug_mode: bool,
 }
 
 /// Construction and access methods.
@@ -169,7 +172,13 @@ impl<R: FeltRng> Client<R> {
             rpc_api,
             tx_executor,
             tx_prover,
+            in_debug_mode,
         }
+    }
+
+    /// Returns true if the client is in debug mode.
+    pub fn is_in_debug_mode(&self) -> bool {
+        self.in_debug_mode
     }
 
     /// Returns a reference to the client's random number generator. This can be used to generate